Removing due to the fact that it was added under the false pretence that it was written by someone it was not.
I had received bits and pieces of this file by a certain member of the internet at large, and put them together. Unbeknownst to me it was the work of another author.

I had gotten the entire code for pipeinbuf from this member, and made some minor modifications to it. For example, using memmove instead of memcpy.

The rest seems similar only because it is the only way to write the code. The file is getting removed to be replaced with a more general copy that can provide more options and functions.

Origional code: http://www.josuttis.com/cppcode/fdstream.html

git-svn-id: http://svn.0x58.com/repos/bsdpanel/0x58-unix/trunk@205 db258d53-41f1-0310-98c4-cc461a13e28c