Updated the license that is included with the source code to the newer OpenBSD style one. Smaller and cleaner and has the same effect as before.
The Event handling has now been cleaned up. I have implemented a roll call type system where events can be enrolled. This way when I finally get around to writing the code for the other event systems (epoll, poll, select, etc ...) and compile time selection of the right ones is enabled, the developer using the library can select which event system to be used, and this way it does not require compile time to decide which one to use, but rather at runtime they all enroll. This also allows the best one for the system to be picked. For example, on FreeBSD you would probably want to use kqueue over select, you also probably want to use epoll over poll on Linux.

Added Event.cc back to the Makefile so that it is now compiled.



git-svn-id: http://svn.0x58.com/repos/bsdpanel/0x58-unix/trunk@220 db258d53-41f1-0310-98c4-cc461a13e28c