Update the license to the cleaner and smaller OpenBSD license. It does the same thing as before, just with less words and thus less space taken up.
This is an event implementation for kqueue for Mac OS X and FreeBSD. As of right now it does not yet return any useful data. I have yet to decide on whether or not I want all callback "classes" to have to derive from one base class, or if I want to do function pointers and let the programmer on the other end deal with keeping certain information around. Function pointers sound pretty good to me though, so I may go that route...



git-svn-id: http://svn.0x58.com/repos/bsdpanel/0x58-unix/trunk@221 db258d53-41f1-0310-98c4-cc461a13e28c