I have changed the name from EventRoll to EventRegistry, while I generally don't like the name Registry for things since it reminds me of the Windows registry and the havoc and hell it wrought on its users.
I am not sure how to accredit parts of the code, so I am hereby doing it in an SVN commit, thanks to Denice
on stackoverflow.com (http://stackoverflow.com/questions/77817/c-runtime-knowledge-of-classes) I was led to 
http://meat.net/2006/03/cpp-runtime-class-registration/ which contained the code that has been mostly used
for the Event Registration.

The original code I had written was similar but this is more elegant and simple.

It will be used along with a EventFactory that is going to come up next, that way as a user of the library
you don't need to worry about what event handling backends are available you just ask the factory to create
one for you!



git-svn-id: http://svn.0x58.com/repos/bsdpanel/0x58-unix/trunk@223 db258d53-41f1-0310-98c4-cc461a13e28c